DEV: Added logging option. Changed print statement to log warning
Changed the warning about array reversal required for function to work
from a basic print statement to a log warning. I decided to post
the notice as logging.warning() due to the fact that the user really
does need to know if their arrays are stored, or being used in reversed
order, unbeknownst to them.