MNT : re-factored snip_method signature
 - merged con_val and iter_num in to single arguements
 - added logic to set default based on spectral_binning is used

This way if user specifies the value of either, that is the value that
will be used (independent of spectral_binning).  If they are not
specified the default values depends on if spectral_binning is used or
not.  The degree of magic number-ness here is not great, but it stays
true to the original code.