DEV: Refactored and revised img arithmetic and logic funcs
There are three major functions included in this commit. The tools are
intended, largely, for use in VisTrails.
These tools are:
arithmetic_basic
arithmetic_custom
logic_basic

arithmetic_basic enables easy manipulation of (a) two arrays, (b) two
constants, or (c) an array and a constant. The arithmetic operations
included in this function are addition, subtraction, multiplication and
division. The only serious restriction included as a part of the
function is the generation of a ValueError if an array or constant is
being divided by 0 (zero).

arithmetic_custom enables more complex arithmetic using up to 8 input
arrays or constants. This function allows the user to specify a custom
arithmetic expression using all of the specified input arrays and
constants and any standard python operator. The available operators
include:
the arithmetic operators such as +, /,
the logic operators such as >, <, >=
bitwise arithmetic operators,
etc.
A complete list of the operators that can be used as a part of the
expression is included in the documentation for the arithmetic_custom
function.

logic_basic enables easy logical evaluation of (a) two arrays, (b) two
constants, or (c) an array and a constant. The logical operations
included in this function are and, or, not, xor, nand, nor, subtract.

Test functions for the tools included in mathops.py are included in the
file test_img_proc.py. This file is anticipated to be the sole test file
for the image processing tools being incorporated into skxray.